Clarified notes for JP4 and JP7
JP4 and JP7 are used to bypass the band filter. The band filter
is required for proper (legal) operation, so bypassing it is
only useful for test or in emergencies. If the jumpers are not
installed, they can't be set wrong, so the jumpers are described
as optional. However, if the jumpers are omitted it is necessary
to install a wire in place of JP7, shorting the center pin and
the FLT pin. This wire was not described in the original text.